move host definitions to host module
